requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
django-nose==1.4.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.0.0
Markdown==3.4.1
nose==1.3.7
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.14.0
